Relocated stream interruption utilities
