+++ src/utils/stream_interrupts.py
@@ -1,13 +1,28 @@
-"""In-memory registry for interrupting active streaming requests."""
+"""Stream interrupt registry and persistence utilities."""
 
 import asyncio
+import datetime
 from collections.abc import Callable, Coroutine
 from dataclasses import dataclass, field
 from enum import Enum
 from threading import Lock
-from typing import Any, Optional
+from typing import Any, Optional, cast
 
+from llama_stack_api import OpenAIResponseMessage
+
+from constants import (
+    INTERRUPTED_RESPONSE_MESSAGE,
+    TOPIC_SUMMARY_INTERRUPT_TIMEOUT_SECONDS,
+)
 from log import get_logger
+from models.common.responses.contexts import ResponseGeneratorContext
+from models.common.responses.responses_api_params import ResponsesApiParams
+from models.common.responses.types import ResponseInput
+from models.common.turn_summary import TurnSummary
+from utils.conversations import append_turn_items_to_conversation
+from utils.query import store_query_results, update_conversation_topic_summary
+from utils.responses import get_topic_summary
+from utils.shields import append_turn_to_conversation
 from utils.types import Singleton
 
 logger = get_logger(__name__)
@@ -146,3 +161,209 @@ def get_stream_interrupt_registry() -> StreamInterruptRegistry:
     and overridden in tests via ``app.dependency_overrides``.
     """
     return StreamInterruptRegistry()
+
+
+def deregister_stream(request_id: str) -> None:
+    """Remove a stream from the interrupt registry after completion or cancellation.
+
+    Parameters:
+    ----------
+        request_id: Unique streaming request identifier.
+    """
+    get_stream_interrupt_registry().deregister_stream(request_id)
+
+
+async def background_update_topic_summary(
+    context: ResponseGeneratorContext,
+    model: str,
+) -> None:
+    """Generate topic summary and update DB/cache in the background.
+
+    Runs as a fire-and-forget task after an interrupted turn is persisted.
+    All errors are caught and logged.
+
+    Parameters:
+    ----------
+        context: The response generator context.
+        model: Model identifier used for topic summary generation.
+    """
+    try:
+        topic_summary = await asyncio.wait_for(
+            get_topic_summary(
+                context.query_request.query,
+                context.client,
+                model,
+            ),
+            timeout=TOPIC_SUMMARY_INTERRUPT_TIMEOUT_SECONDS,
+        )
+        if topic_summary:
+            update_conversation_topic_summary(
+                context.conversation_id,
+                topic_summary,
+                user_id=context.user_id,
+                skip_userid_check=context.skip_userid_check,
+            )
+    except asyncio.TimeoutError:
+        logger.warning(
+            "Topic summary timed out for interrupted turn, request %s",
+            context.request_id,
+        )
+    except Exception:  # pylint: disable=broad-except
+        logger.exception(
+            "Failed to generate topic summary for interrupted turn, request %s",
+            context.request_id,
+        )
+
+
+async def persist_interrupted_turn(
+    context: ResponseGeneratorContext,
+    responses_params: ResponsesApiParams,
+    turn_summary: TurnSummary,
+    background_topic_summary_tasks: list[asyncio.Task[None]],
+    original_input: Optional[ResponseInput] = None,
+) -> None:
+    """Persist the user query and an interrupted response into the conversation.
+
+    Called when a streaming request is cancelled so the exchange is not lost.
+    Persists immediately with topic_summary=None so the conversation exists
+    when the client fetches. Topic summary is generated in a background task
+    and updated when ready.
+
+    Parameters:
+    ----------
+        context: The response generator context.
+        responses_params: The Responses API parameters.
+        turn_summary: TurnSummary with llm_response already set to the
+            interrupted message.
+        background_topic_summary_tasks: Mutable list tracking fire-and-forget
+            topic summary tasks for graceful shutdown.
+        original_input: In compacted mode, the original user input before the
+            explicit-input rewrite. When set, the turn is persisted against it
+            (the ``conversation`` parameter was dropped, and
+            ``responses_params.input`` is the explicit rewrite); ``None``
+            otherwise (LCORE-1572).
+    """
+    try:
+        if original_input is not None:
+            await append_turn_items_to_conversation(
+                context.client,
+                responses_params.conversation,
+                original_input,
+                [
+                    OpenAIResponseMessage(
+                        role="assistant", content=INTERRUPTED_RESPONSE_MESSAGE
+                    )
+                ],
+            )
+        else:
+            await append_turn_to_conversation(
+                context.client,
+                responses_params.conversation,
+                cast(str, responses_params.input),
+                INTERRUPTED_RESPONSE_MESSAGE,
+            )
+    except Exception:  # pylint: disable=broad-except
+        logger.exception(
+            "Failed to append interrupted turn to conversation for request %s",
+            context.request_id,
+        )
+
+    try:
+        completed_at = datetime.datetime.now(datetime.UTC).strftime(
+            "%Y-%m-%dT%H:%M:%SZ"
+        )
+        store_query_results(
+            user_id=context.user_id,
+            conversation_id=context.conversation_id,
+            model=responses_params.model,
+            completed_at=completed_at,
+            started_at=context.started_at,
+            summary=turn_summary,
+            query=context.query_request.query,
+            skip_userid_check=context.skip_userid_check,
+            topic_summary=None,
+        )
+
+        if (
+            not context.query_request.conversation_id
+            and context.query_request.generate_topic_summary
+        ):
+            task = asyncio.create_task(
+                background_update_topic_summary(
+                    context=context,
+                    model=responses_params.model,
+                )
+            )
+            background_topic_summary_tasks.append(task)
+            task.add_done_callback(background_topic_summary_tasks.remove)
+    except Exception:  # pylint: disable=broad-except
+        logger.exception(
+            "Failed to store interrupted query results for request %s",
+            context.request_id,
+        )
+
+
+def register_interrupt_callback(
+    context: ResponseGeneratorContext,
+    responses_params: ResponsesApiParams,
+    turn_summary: TurnSummary,
+    background_topic_summary_tasks: list[asyncio.Task[None]],
+    original_input: Optional[ResponseInput] = None,
+) -> list[bool]:
+    """Build an interrupt callback and register the stream for cancellation.
+
+    The callback is invoked by ``cancel_stream`` when the client
+    interrupts, so persistence runs regardless of where the
+    ``CancelledError`` is raised in the ASGI stack.
+
+    A mutable one-element list is used as a shared guard so the
+    callback and the in-generator ``CancelledError`` handler never
+    both persist the same turn.
+
+    Parameters:
+    ----------
+        context: The response generator context.
+        responses_params: The Responses API parameters.
+        turn_summary: TurnSummary populated during streaming.
+        background_topic_summary_tasks: Mutable list tracking fire-and-forget
+            topic summary tasks for graceful shutdown.
+        original_input: In compacted mode, the original user input before the
+            explicit-input rewrite; ``None`` otherwise.
+
+    Returns:
+    -------
+        A mutable list ``[False]`` used as a persist-done guard; the
+        caller should check ``guard[0]`` before persisting and set
+        it to ``True`` afterwards.
+    """
+    guard: list[bool] = [False]
+
+    async def _on_interrupt() -> None:
+        if guard[0]:
+            return
+        guard[0] = True
+        turn_summary.llm_response = INTERRUPTED_RESPONSE_MESSAGE
+        await persist_interrupted_turn(
+            context,
+            responses_params,
+            turn_summary,
+            background_topic_summary_tasks,
+            original_input,
+        )
+
+    current_task = asyncio.current_task()
+    if current_task is not None:
+        get_stream_interrupt_registry().register_stream(
+            request_id=context.request_id,
+            user_id=context.user_id,
+            task=current_task,
+            on_interrupt=_on_interrupt,
+        )
+    else:
+        logger.warning(
+            "No current asyncio task for request %s; "
+            "stream interruption will not be available",
+            context.request_id,
+        )
+
+    return guard
